Fix styling for 03-flex-header-2
Added styling to match the requirements. In the readme a solution using additional divs to align the left and right parts of the header is suggested. The same result can be achieved by using auto right margin on the links class

For completeness, added a second header to the html with its own classes to achieve the same visual result by using the additional divs method as suggested in the readme